requirements.txt
allure-behave
allure-pytest
pytest
behave
python-dotenv
selenium